Area Friendly target handler update
Static comparator to prevent initialization on each call.
List is created after single-target checks are performed.
Using LinkedList instead ArrayList due resize overhead.
Adding target only if it's not null.
Sorting the list of creatures after it's complete, not after adding the
target (that wouldn't have an effect on the list since it has a single
element).
Performing maximum target check first.
Skipping hidden creatures.
Skipping, when in siege, creatures that are not in the same siege side.