Syntax: Fixing illegal FlagWar.super expression.
Using super to get the this reference as a reference of the type of the
super class is not valid. According to java documentation super can be
used to access members and methods of it's super class. Using it to get
a reference to the current class is not documented. In the past ECJ
allowed this expression and hapilly compiled. But tests one year ago has
shown that the JDK compiler in tools.jar was reporting it as an error.
Now as reported @Sacrifice, the last update of Eclipse also complains
about this, so this expression is now removed and the code a little bit
cleaned using a lambda instead of inner class.

Reported By: @Sacrifice
Patch By: @Sacrifice, @HorridoJoho